+++ src/Product.py
@@ -1,0 +1,15 @@
+class Product:
+    """
+    Класс продукта
+    """
+
+    name: str
+    description: str
+    price: float
+    quantity: int
+
+    def __init__(self, name, description, price, quantity):
+        self.name = name
+        self.description = description
+        self.price = price
+        self.quantity = quantity
